Add gravity directly to velocity.
The paper [1] recommended doing it in two steps, but this seems to
have a problem: when colliding with things above the circle, gravity
moves down by two pixels and it doesn't look like the collision was
flush. The problem that the paper sites with adding gravity directly
to velocity doesn't seem to matter.

[1] http://www.paulnettle.com/pub/FluidStudios/CollisionDetection/Fluid_Studios_Generic_Collision_Detection_for_Games_Using_Ellipsoids.pdf